Moves nearly all ScriptingProperties to Mods.Common
