+++ OpenRA.Mods.Common/Scripting/Properties/GeneralProperties.cs
@@ -11,12 +11,10 @@
 using Eluant;
 using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Scripting
+namespace OpenRA.Mods.Common.Scripting
 {
 	[ExposedForDestroyedActors]
 	[ScriptPropertyGroup("General")]
