+++ OpenRA.Mods.Common/Scripting/Properties/MobileProperties.cs
@@ -10,12 +10,10 @@
 
 using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Scripting
+namespace OpenRA.Mods.Common.Scripting
 {
 	[ScriptPropertyGroup("Movement")]
 	public class MobileProperties : ScriptActorProperties, Requires<MobileInfo>
