+++ OpenRA.Mods.Common/Scripting/Properties/PlaneProperties.cs
@@ -10,11 +10,10 @@
 
 using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Scripting
+namespace OpenRA.Mods.Common.Scripting
 {
 	[ScriptPropertyGroup("Movement")]
 	public class PlaneProperties : ScriptActorProperties, Requires<PlaneInfo>
