+++ OpenRA.Mods.Common/Scripting/Properties/RepairableBuildingProperties.cs
@@ -12,7 +12,7 @@ using OpenRA.Mods.Common.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Scripting
+namespace OpenRA.Mods.Common.Scripting
 {
 	[ScriptPropertyGroup("General")]
 	public class RepairableBuildingProperties : ScriptActorProperties, Requires<RepairableBuildingInfo>
